refactor: split workflow.go into focused files
Split internal/interactive/workflow.go (528 lines) into 5 files based on
responsibility:

- workflow_step.go (27 lines) - WorkflowStep type and String method
- workflow.go (74 lines) - Workflow type and basic operations
- workflow_manager.go (227 lines) - WorkflowManager and multi-workflow lifecycle
- workflow_executor.go (74 lines) - WorkflowExecutor and execution logic
- workflow_placeholders.go (150 lines) - Placeholder resolution and interactive input

This follows the same pattern as previous splits (input.go, keys.go, ui.go)
to improve code organization and maintainability. All existing tests pass
without modification.

Changes from code review:
- Remove redundant len(parts) == 0 check
- Improve mutex handling in CloneWorkflow
- Add empty input validation in placeholder resolution
